Move all test code into one file.
